backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.2.0
python-multipart==0.0.6
requests>=2.31.0
psycopg2-binary